An alternative to #17928. @petermattis let me know if you'd like me to test this. As I mentioned, all I've done so far is observe a slight slowdown in `BenchmarkIterOnEngine`:

```
name                                 old time/op  new time/op  delta
IterOnEngine_RocksDB/writes=10-8     2.08µs ± 5%  2.13µs ± 2%    ~     (p=0.548 n=5+5)
IterOnEngine_RocksDB/writes=100-8    2.22µs ± 4%  2.26µs ± 4%    ~     (p=0.548 n=5+5)
IterOnEngine_RocksDB/writes=1000-8   2.27µs ± 1%  2.46µs ± 4%  +8.41%  (p=0.008 n=5+5)
IterOnEngine_RocksDB/writes=10000-8  2.61µs ± 5%  2.66µs ± 2%    ~     (p=0.421 n=5+5)
```

---

Previously, if we had a large-enough open file ulimit, we'd configure
RocksDB without an open file limit (`max_open_files = -1`). This
configuration enabled a fast path where RocksDB could avoid an lookup in
its open files cache, since it could assume that every opened file would
stay open forever.

Unfortunately, a 2TB restore creates tens of thousands of 2MB SSTs,
quickly exceeding the open file ulimit of 10k. With `max_open_files = -1`
the call to `open` that exceeds the limit crashes the server.

Instead, set the `max_open_files` to the configured ulimit so that
RocksDB evicts a file that hasn't been recently accessed instead of
crashing the server. This introduces two performance hits: an obvious
hit, where accessing restored data is slow because the necessary files
have been evicted from the open files cache, and a less-obvious hit,
where every operation now requires an open-file cache lookup, even when
the number of SSTs is well below the open file limit.