@danhhz 2017-08-28
We also should probably account somehow for files that are opened outside of rocksdb (AddSSTable sideloading for example)