@danhhz 2017-08-28
I haven't looked at how the rocksdb max files limit is implemented, but if it leaves open exactly the number you request and that's how many the user has ulimited, then even one extra will put you over.

It may end up that we don't have to do anything, I just want to make sure it's thought through