@petermattis 2017-08-28
> I haven't looked at how the rocksdb max files limit is implemented, but if it leaves open exactly the number you request and that's how many the user has ulimited, then even one extra will put you over.

Right, so that's why we specify `max_open_files` as somewhat less than the open file limit. We already do this to allow some file descriptors to be used for sockets and log files.