@petermattis 2017-08-28
Btw, this can be merged before #17959. I'd like to know what performance we're leaving on the table by setting `max_open_files != -1`, but clearly it is better to be somewhat slower than to crash.