@bdarnell 2017-08-28
Review status: 0 of 3 files reviewed at latest revision, 3 unresolved discussions, some commit checks failed.

---

*[pkg/server/config_unix.go, line 35 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-KsdtP6TgQs2yrHYS2H4:-KsdtP6TgQs2yrHYS2H5:b-c9q3wi) ([raw file](https://github.com/cockroachdb/cockroach/blob/ee9cf6e944de91d1f65e3359f2c1077f9eb5b6bd/pkg/server/config_unix.go#L35)):*
> ```Go
> }
> 
> func setOpenFileLimitInner(physicalStoreCount int) (uint64, error) {
> ```

This signature also needs to be changed in config_windows.go.

---

*[pkg/server/config_unix.go, line 99 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-Ksdt88cwLBfPgIs2dS3:-Ksdt88cwLBfPgIs2dS4:bgx2nog) ([raw file](https://github.com/cockroachdb/cockroach/blob/ee9cf6e944de91d1f65e3359f2c1077f9eb5b6bd/pkg/server/config_unix.go#L99)):*
> ```Go
> 	// value, we can use the current limit.
> 	if rLimit.Cur >= recommendedOpenFileLimit {
> 		return rLimit.Cur, nil
> ```

We don't want to use the whole limit here, we want something like `rLimit.Cur - recommendedNetworkFileDescriptors`. Originally, the idea was that engine.RecommendedMaxOpenFiles would be an upper limit on the number of file descriptors we'd use and any excess would go to sockets. I'm concerned about not having enough file descriptors available for sockets; maybe we should split the difference and divide any file descriptors over the "recommended" levels between files and sockets.

---

*[pkg/storage/engine/rocksdb.go, line 90 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-Ksdu5SGfYcWWMmM2XrL:-Ksdu5SGfYcWWMmM2XrM:bildcc9) ([raw file](https://github.com/cockroachdb/cockroach/blob/ee9cf6e944de91d1f65e3359f2c1077f9eb5b6bd/pkg/storage/engine/rocksdb.go#L90)):*
> ```Go
> 	// RecommendedMaxOpenFiles is the recommended value for rocksDB's
> 	// max_open_files option. If more file descriptors are available than the
> 	// recommended number, than the default value is used.
> ```

This comment needs updating. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958)*
<!-- Sent from Reviewable.io -->