@benesch 2017-08-28
Review status: 0 of 4 files reviewed at latest revision, 3 unresolved discussions, some commit checks pending.

---

*[pkg/server/config_unix.go, line 35 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-KsdtP6TgQs2yrHYS2H4:-KsdyXLRP_ZXLkZ4dSQo:bjqnukk) ([raw file](https://github.com/cockroachdb/cockroach/blob/ee9cf6e944de91d1f65e3359f2c1077f9eb5b6bd/pkg/server/config_unix.go#L35)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

This signature also needs to be changed in config_windows.go.
</blockquote></details>

Ooh, good catch. Thanks.

---

*[pkg/server/config_unix.go, line 99 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-Ksdt88cwLBfPgIs2dS3:-Ksdyb50_HS8y0d8Dsel:b-39annp) ([raw file](https://github.com/cockroachdb/cockroach/blob/ee9cf6e944de91d1f65e3359f2c1077f9eb5b6bd/pkg/server/config_unix.go#L99)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

We don't want to use the whole limit here, we want something like `rLimit.Cur - recommendedNetworkFileDescriptors`. Originally, the idea was that engine.RecommendedMaxOpenFiles would be an upper limit on the number of file descriptors we'd use and any excess would go to sockets. I'm concerned about not having enough file descriptors available for sockets; maybe we should split the difference and divide any file descriptors over the "recommended" levels between files and sockets.
</blockquote></details>

Whoops, looks like this also needs to divide by the number of stores. Did the simple solution:

```
(rLimit.Cur - recommendedNetworkFileDescriptors) / uint64(physicalStoreCount)
```

I held off on doing anything more complicated to allocate fds between network sockets and stores, since this is strictly better than what we had before (where RocksDB could completely starve us of file descriptors).

---

*[pkg/storage/engine/rocksdb.go, line 90 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-Ksdu5SGfYcWWMmM2XrL:-KsdyRp_hyZXCLECyFUX:b-dtkdrq) ([raw file](https://github.com/cockroachdb/cockroach/blob/ee9cf6e944de91d1f65e3359f2c1077f9eb5b6bd/pkg/storage/engine/rocksdb.go#L90)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

This comment needs updating. 
</blockquote></details>

Done, thanks.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958)*
<!-- Sent from Reviewable.io -->