@bdarnell 2017-08-28
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 3 files at r1, 3 of 3 files at r2.
Review status: all files reviewed at latest revision, 4 unresolved discussions, some commit checks pending.

---

*[pkg/server/config_unix.go, line 123 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-Kse-qxJuIcT-9yyY7b7:-Kse-qxJuIcT-9yyY7b8:b-hqlfc3) ([raw file](https://github.com/cockroachdb/cockroach/blob/4577642b50f214987255b8cca7ff75a8d9442523/pkg/server/config_unix.go#L123)):*
> ```Go
> 	// Always sacrifice all but the minimum needed network descriptors to be
> 	// used by the stores.
> 	return rLimit.Cur - minimumNetworkFileDescriptors/uint64(physicalStoreCount), nil
> ```

The parentheses still need to be here for mathematical precedence. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-:-Kse02YO_2d3UZKXVwEJ:bnfp4nl)*
<!-- Sent from Reviewable.io -->