@benesch 2017-08-28
Review status: 3 of 4 files reviewed at latest revision, 4 unresolved discussions.

---

*[pkg/server/config_unix.go, line 123 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-Kse-qxJuIcT-9yyY7b7:-Kse1aOt1sOmeon8XT_D:b-n30ulx) ([raw file](https://github.com/cockroachdb/cockroach/blob/4577642b50f214987255b8cca7ff75a8d9442523/pkg/server/config_unix.go#L123)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

The parentheses still need to be here for mathematical precedence. 
</blockquote></details>

Eek, that they do. Thanks. :man_facepalming: 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958)*
<!-- Sent from Reviewable.io -->