@BramGruneir 2017-08-28
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---







*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17958#-:-Kse8vIRUtmkj9X7NwCf:bnfp4nl)*
<!-- Sent from Reviewable.io -->